+++ freestream/__init__.py
@@ -1,0 +1,1 @@
+from .pages.utils.styles import footer
