+++ freestream/pages/2_🖼️_Image_Upscaler.py
@@ -1,7 +1,7 @@
 import os
 
 import streamlit as st
-from pages.utils import image_upscaler, footer
+from pages import image_upscaler, footer
 
 # Set expandable_segments
 os.environ["PYTORCH_CUDA_ALLOC_CONF"] = "expandable_segments:True"
