+++ freestream/🏡_Home.py
@@ -1,6 +1,6 @@
 import streamlit as st
 
-from freestream.pages.utils import footer
+from freestream import footer
 
 st.set_page_config(page_title="FreeStream: Unlimited Access to AI Tools", page_icon="🏡")
